Backend/requirement.txt
fastapi[standard]
uvicorn[standard]
PyJWT
passlib[bcrypt]
python-multipart